Updated setup to be more robust
Altered version to not require self import, simplified commands to
add additional, and updated the package directory structure to be
more robust for installing additional packages within the src folder
with discovery.  Also added a manifest file so that it worked
appropriately across distributions.

This fixes things so that they work in windows from a bare pip install.